OFED 23.10-0.5.5.0

+++ drivers/net/ethernet/mellanox/mlx5/core/sf/vhca_event.h
@@ -8,7 +8,7 @@
 
 struct mlx5_vhca_state_event {
 	u16 function_id;
-	u32 sw_function_id;
+	u16 sw_function_id;
 	u8 new_vhca_state;
 };
